test: changed Tasks component name and added tests
