+++ src/components/index.js
@@ -2,4 +2,4 @@ export { default as About } from './About';
 export { default as AddTask } from './AddTask';
 export { default as Footer } from './Footer';
 export { default as Header } from './Header';
-export { default as Tasks } from './Tasks';
+export { default as TaskList } from './TaskList';
